- mutation extended
- static cell bytes do not need to be multiple of 3 in case of cell computers

+++ source/ModelGpu/Base.cuh
@@ -52,7 +52,8 @@ public:
         checkCudaErrors(cudaMemcpy(_array, randomNumbers.data(), sizeof(int)*size, cudaMemcpyHostToDevice));
     }
 
-    __device__ __inline__ float random(int maxVal)
+
+    __device__ __inline__ int random(int maxVal)
     {
         int number = getRandomNumber();
         return number % (maxVal + 1);
